Revert "Inventory test and simplex improvement finished"

+++ Alien-Logic/src/main/java/icesi/vip/alien/service/inventoryManagement/ContinuousRevSS.java
@@ -2,27 +2,27 @@ package icesi.vip.alien.service.inventoryManagement;
 
 public class ContinuousRevSS implements InventorySystem{
 
-	private int maxLevelInventory;
-	private int minLevelInventory;
+	private float maxLevelInventory;
+	private float minLevelInventory;
 	
 	@Override
 	public double calculateQuantity() {
 			return maxLevelInventory - minLevelInventory;
 	}
 
-	public int getMaxLevelInventory() {
+	public float getMaxLevelInventory() {
 		return maxLevelInventory;
 	}
 
-	public void setMaxLevelInventory(int maxLevelInventory) {
+	public void setMaxLevelInventory(float maxLevelInventory) {
 		this.maxLevelInventory = maxLevelInventory;
 	}
 
-	public int getMinLevelInventory() {
+	public float getMinLevelInventory() {
 		return minLevelInventory;
 	}
 
-	public void setMinLevelInventory(int minLevelInventory) {
+	public void setMinLevelInventory(float minLevelInventory) {
 		this.minLevelInventory = minLevelInventory;
 	}
